+++ NEWS.md
@@ -1,6 +1,6 @@
 # dev
 
-* Modified the default vertical positioning of column and panel headings in forest_plot()
+* Modified the default vertical positioning of column and panel headings, and x-axis label, in forest_plot()
 * Added argument panel.headings.align to forest_plot() to control centring of panel headings.
 * Added argument col.heading.rule to forest_plot() to add a horizontal rule below column headings.
 * Added argument axis.title.margin to theme_ckb() and ckb_style().